+++ gulpfile.js
@@ -221,6 +221,8 @@ async function wipe() {
 		del([ 'static/html/*' ]),
 		del([ 'static/json/*' ]),
 		del([ 'static/banner/*' ]),
+		del([ 'static/flag/*' ]),
+		del([ 'static/asset/*' ]),
 		del([ 'static/css/*' ]),
 	]);
 
@@ -352,6 +354,7 @@ async function custompages() {
 	])
 	.pipe(gulppug({
 		locals: {
+			authLevelNames: ['Admin', 'Global Staff', 'Global Board Owner', 'Global Board Mod', 'Regular User'],
 			early404Fraction: config.get.early404Fraction,
 			early404Replies: config.get.early404Replies,
 			meta: config.get.meta,